// src/app/pages/Leader/model/community.ts
export interface CommunityInfo {
  id: string;
  clientName: string;
  apiKey: string;
  gmail: string;
  googleCredentialPassword: string;
  apiSecret: string;
  isPublic: boolean;
  usersCount: number;
  logoUrl: string;
  facebookLink: string;
  startDate: string;
  endDate: string;
  isRequiredTraineeNationalId: boolean;
  isRequiredTraineePhoneNumber: boolean;
}